+++ app/build.gradle
@@ -11,7 +11,7 @@ android {
     }
 
     compileSdkVersion 28
-    buildToolsVersion '28.0.2'
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "ru.vyaacheslav.suhov.imeit_rsp"
         minSdkVersion 19
@@ -34,15 +34,6 @@ android {
         skipDeprecated = true
         reportUndocumented = true
     }
-    
-    task deleteUnusedDokkaAssets(type: Delete) {
-        file("$rootDir/docs").listFiles().each {
-            if( it.name.contains("android.R" ) ) projects.delete it
-            if( it.name.contains("android.support" ) ) projects.delete it
-            if( it.name.contains("com.google" ) ) projects.delete it
-            if( it.name.contains("com.crashlytics" ) ) projects.delete it
-        }
-    }
 }
 
 dependencies {
@@ -52,8 +43,9 @@ dependencies {
     implementation 'com.android.support:design:28.0.0'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation "org.jetbrains.anko:anko-commons:$anko_version"
-    implementation 'com.google.android.gms:play-services-maps:15.0.1'
+    implementation 'com.google.android.gms:play-services-maps:16.0.0'
     implementation 'com.android.support:recyclerview-v7:28.0.0'
+    implementation project(':core')
 }
 
 repositories {
