+++ app/src/main/java/ru/vyaacheslav/suhov/imeit/adapters/SettingsBellsAdapter.kt
@@ -10,11 +10,11 @@ import android.widget.CheckBox
 import android.widget.EditText
 import android.widget.TextView
 import ru.vyaacheslav.suhov.imeit.R
-import ru.vyaacheslav.suhov.imeit.core.ui.SettingsPoly
+import ru.vyaacheslav.suhov.imeit.objects.SettingsPoly
 
 class SettingsBellsAdapter(
         private val context: Context,
-        private val list: List<SettingsPoly>
+       private val list: List<SettingsPoly>
 
 ) : RecyclerView.Adapter<SettingsBellsAdapter.ViewHolder>() {
 
@@ -44,13 +44,14 @@ class SettingsBellsAdapter(
     @SuppressLint("SetTextI18n")
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
 
-        holder.name.text = list[position].name
-        holder.value.setText(list[position].data.toString(), TextView.BufferType.EDITABLE)
+      /*  holder.name.text = list[position].name
+        holder.value.setText(list[position].data.toString(), TextView.BufferType.EDITABLE)*/
     }
 
     // Возвращает размер данных (вызывается layout manager-ом)
     override fun getItemCount(): Int {
-        return list.size
+      /*  return list.size*/
+        return 0
     }
