--- app/src/main/java/ru/vyaacheslav/suhov/imeit/core/CountBells.kt
@@ -1,6 +1,0 @@
-package ru.vyaacheslav.suhov.imeit.core
-data class CountBells(var num: String,
-                      var topStr: String,
-                      var botStr: String,
-                      var topOut: String,
-                      var botOut: String)
