+++ app/src/main/java/ru/vyaacheslav/suhov/imeit/objects/SettingsPoly.kt
@@ -1,4 +1,4 @@
-package ru.vyaacheslav.suhov.imeit.core.ui
+package ru.vyaacheslav.suhov.imeit.objects
 
 data class SettingsPoly(
