+++ core/src/main/java/core/EducationalProcess.kt
@@ -1,8 +1,7 @@
-package ru.vyaacheslav.suhov.imeit.core
+package core
 
+import core.objects.BellCount
 import java.util.*
-
-//TODO: Данный клас должен быть в отдельном модуле
 
 /** Данный класс реализует полную работу с времннными диапазонами учебного процесса :з
  *  @author Sukhov Viacheslav
@@ -14,7 +13,7 @@ class EducationalProcess {
      *  @param table - Обязательное поле которое принимает
      *  @return Вернет всю статистику у виде ArrayList<Int> данные в минутах*/
 
-    fun getAllStatistic(table: List<CountBells>): Array<Int> {
+    fun getAllStatistic(table: List<BellCount>): Array<Int> {
 
         /** Возвращаем полученную статистику */
         return getPairInfo()
