+++ core/src/main/java/core/objects/BellCount.kt
@@ -1,0 +1,12 @@
+package core.objects
+
+/**@author SukhovViacheslav
+ * @version 1.0
+ *
+ * This object a table is for android, web and others. */
+
+data class BellCount(var num: String,
+                     var topStr: String,
+                     var botStr: String,
+                     var topOut: String,
+                     var botOut: String)
