+++ core/src/main/java/core/objects/application/AppSetup.kt
@@ -1,0 +1,6 @@
+package core.objects.application
+
+data class AppSetup (
+        val locale:String = "ru",
+        val timeFormat:String = "standard"
+)
